requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-iot-device==2.12.0
certifi==2023.5.7
charset-normalizer==3.1.0
deprecation==2.1.0
frozenlist==1.3.3
idna==3.4
janus==1.0.0
multidict==6.0.4
packaging==23.1
paho-mqtt==1.6.1
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-unixsocket==0.3.0
typing_extensions==4.6.1
urllib3==1.26.16
yarl==1.9.2
